I should compile more times :)
git-svn-id: http://svn.l2jfree.com/l2jfree@2282 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_IL/src/main/java/net/sf/l2j/gameserver/model/L2Character.java
@@ -5660,9 +5660,9 @@ public abstract class L2Character extends L2Object
      * Return True if the target is front L2Character and can be seen.<BR><BR>
      * degrees = 0..180, front->sides->back
      */
-    public boolean isInFront(L2Character target, int degrees)
+    public boolean isInFront(L2Character target, double degrees)
     {
-        int head = getHeadingTo(target, false);
+        double head = getHeadingTo(target, false);
         if (head <= 32768*degrees/180 || head >= (65536 - 32768*degrees/180))
             return true;
         return false;
